Moved the AudioChecker to the Proper Class
